Add solutions file to What's Cooking?
This problem uses a training set where the predictors are a list of ingredients used in a recipe, and the outcome variable is the cuisine of the recipe (from a list of 20 possible cuisines). To goal is to identify the cuisines in a test set of recipes as accurately as possible.